requirements/common.txt
APScheduler==3.5.1
Django==2.0.7
django-allauth==0.36.0
django-braces==1.13.0
django-crispy-forms==1.7.2
django-jsonfield==1.0.1
django-simple-history==2.2.0
djangorestframework==3.8.2
django-filter==1.1.0
django-datetime-widget==0.9.3
django-datatables-view==1.16.0
iso8601==0.1.11
psycopg2-binary==2.7.5
pytz==2018.5
rabbitmq-admin==0.2
SQLAlchemy==1.2.9

# Pinning to correct version for sfm-utils
requests==2.9.1
urllib3==1.12

# Testing
mock==1.3.0
